feat: add stats frame
